Migrate from SGE to toolforge-jobs
This commit also:
- Installs pywikibot from pypi instead of git
- Adds a script to construct the virtual environment
- Runs the bot on Python 3.11

Unfortunately toolforge-jobs does not have support for
ConcurrencyPolicy: Replace or activeDeadlineSeconds (T306391),
so there is still a potential for stuck jobs to stop the bot from
running. K8s should still be more reliable than the grid, so I don't
know how much of a problem this will be. If it remains a problem, it
could be fixed with a custom CronJob or adding a timeout to the wrapper
script.

Bug: T319848
Change-Id: I27c91a25b8fe2b78595b04d22ffebdb464b90b1f