Skip edit if server responds with HTTP 50x
Follows-up f0d0b20e3e, which added skipping for RuntimeException.
This took care of cases where api.php responds with a MW API error
code.

It did not take care of cases where the request fails at the CDN edge
before it even reaches MediaWiki, because those return HTTP 501-503,
with an HTML response instead of a JSON response, which means we
throw JsonException, which our skip logic does not tolerate and consider
a fatal error. I'd like to keep that as a fatal error.

Instead, detect the HTTP error before we parse, and throw
RuntimeException for it.

Change-Id: Ifbe9fc4ff66cbe437f8cd0a536e76abf34a696b4