phan: Add missing phan config file, and upgrade Phan
In the refactor I forgot to commit my config file it seems.

Phan on PHP <= 8.2 gracefully notes this fact without exiting non-zero:

```
$ composer test
> parallel-lint
…
> phpunit
…
> phan --allow-polyfill-parser
…
Phan did not parse any files in the project "labs-tools-fileprotectionsync" - This may be an issue with the Phan config or CLI options.
> minus-x check .
…
[exit=0] 1s $
```

On PHP 8.4, however, this code paths happens trigger an unrelated
internal deprecation and so fails WMF CI at
https://gerrit.wikimedia.org/r/1149847

```
Phan\Issue::__invoke(): Implicitly marking parameter $suggestion as
nullable is deprecated, the explicit nullable type must be used instead
```

... which is fixed by updating to a newer Phan, but drew me to the
above mistake, so this adds the config file, updates Phan, and makes
it pass.

Change-Id: I869615d05f978040ff3ad498f040b94721f5ff7e